backend/pyproject.toml
[project]
name = "voiceagent-backend"
version = "1.0.0"
description = "Voice Agent API - FastAPI backend with WebSocket support"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web stack
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "websockets>=12.0",
    "python-multipart>=0.0.6",
    "httpx>=0.25.0",
    # Voiceagent runtime (so backend venv is self-contained)
    "openai-agents[voice]>=0.1.0",
    "groq>=0.11.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "numpy>=1.24.0",
    "rich>=13.0.0",
    "scipy>=1.10.0",
    "alembic>=1.18.4",
]

[project.scripts]
app = "index:app"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
local-audio = [
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
]

[tool.uv]
package = false
